# Make 'gmaven.bzl' file generating more faster