--- a/GMavenToBazel.java
+++ b/GMavenToBazel.java
@@ -12,7 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.import java.io.FileWriter;
 
+
+import java.io.BufferedInputStream;
+import java.io.BufferedWriter;
 import java.io.FileWriter;
+import java.io.InputStream;
 import java.io.PrintWriter;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -22,13 +26,16 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
+import java.util.stream.Collectors;
+
 import javax.xml.parsers.DocumentBuilderFactory;
+
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
-
 /**
  * A command line tool for generating Bazel workspace rules for all JARs and AARs on
  * https://maven.google.com.
@@ -57,44 +64,51 @@ public class GMavenToBazel {
 
   public static void main(String[] args) throws Exception {
     Document masterIndex = parseUrl(new URL(MASTER_INDEX_URL));
-    List<String> groupIds = getGroupdIds(masterIndex);
-    Map<String, String> repositoryNameToRuleType = new HashMap<>();
-    Map<String, String> repositoryNameToTargetName = new HashMap<>();
-    Map<String, String> repositoryNameToUrl = new HashMap<>();
-    Map<String, Set<String>> repositoryNameToRepositoryNameDeps = new HashMap<>();
-    for (String groupId : groupIds) {
-      Map<String, Set<String>> artifactsToVersions = getArtifactsToVersions(groupId);
-      for (String artifactId : artifactsToVersions.keySet()) {
-        for (String version : artifactsToVersions.get(artifactId)) {
-          Document pom = parseUrl(getPomUrl(groupId, artifactId, version));
-          Element project = (Element) pom.getElementsByTagName("project").item(0);
-          String packaging = getPackaging(project);
-          if (!PACKAGING_TO_RULE.containsKey(packaging)) {
+    Map<String, Artifact> artifacts = getGroupdIds(masterIndex)
+        .parallelStream()
+        .flatMap(groupId -> {
+          Map<String, Set<String>> artifactsToVersions = getArtifactsToVersions(groupId);
+          return artifactsToVersions.entrySet()
+              .stream()
+              .flatMap(entry -> {
+                String artifactId = entry.getKey();
+                Set<String> versions = entry.getValue();
+                return versions.stream().map(version -> new String[]{groupId, artifactId, version});
+              });
+        })
+        .map(ids -> {
+          String groupId = ids[0], artifactId = ids[1], version = ids[2];
+          try {
+            Document pom = parseUrl(getPomUrl(groupId, artifactId, version));
+
+            String packaging = getPackaging(pom);
             // We don't support APK packaging yet.
-            continue;
+            if (!PACKAGING_TO_RULE.containsKey(packaging)) {
+              return null;
+            }
+            Artifact artifact = new Artifact(groupId, artifactId, version, packaging);
+            artifact.dependencies = getDependencyRepositoryNames(pom);
+
+            return artifact;
+          } catch (Exception e) {
+            throw new RuntimeException(e);
           }
-          String repositoryName = getRepositoryName(groupId, artifactId, version);
-          repositoryNameToRuleType.put(repositoryName, PACKAGING_TO_RULE.get(packaging));
-          repositoryNameToTargetName.put(
-              repositoryName, getTargetName(groupId, artifactId, version, packaging));
-          repositoryNameToUrl.put(repositoryName, getUrl(groupId, artifactId, version, packaging));
-          repositoryNameToRepositoryNameDeps.put(repositoryName, getDependencyRepositoryNames(pom));
-        }
-      }
-    }
+        }).filter(Objects::nonNull)
+        .collect(Collectors.toMap(artifact -> artifact.repositoryName, artifact -> artifact));
 
-
-    PrintWriter bzlWriter = new PrintWriter(new FileWriter(OUTPUT_FILE));
+    PrintWriter bzlWriter = new PrintWriter(new BufferedWriter(new FileWriter(OUTPUT_FILE), 32 * 1024));
     bzlWriter.println("load('//:import_external.bzl', 'java_import_external', 'aar_import_external')");
     bzlWriter.println("def gmaven_rules():");
-    for (String repositoryName : repositoryNameToRuleType.keySet()) {
-      String ruleType = repositoryNameToRuleType.get(repositoryName);
-      String url = repositoryNameToUrl.get(repositoryName);
+
+    for (String repositoryName : artifacts.keySet()) {
+      final Artifact artifact = artifacts.get(repositoryName);
+      final String url = artifact.toArtifactUrl();
+      final String ruleType = artifact.rule();
       bzlWriter.println();
       bzlWriter.println(String.format("  %s(", ruleType));
       bzlWriter.println(String.format("      name = '%s',", repositoryName));
       bzlWriter.println("      licenses = ['notice'], # apache");
-      if (ruleType == "aar_import_external") {
+      if ("aar_import_external".equals(ruleType)) {
           bzlWriter.println(String.format("      aar_urls = ['%s'],", url));
           bzlWriter.println("      aar_sha256 = '',");
       } else {
@@ -102,14 +116,13 @@ public class GMavenToBazel {
           bzlWriter.println("      jar_sha256 = '',");
       }
       bzlWriter.println("      deps = [");
-      for (String repositoryNameDep : repositoryNameToRepositoryNameDeps.get(repositoryName)) {
-        String targetNameDep = repositoryNameToTargetName.get(repositoryNameDep);
+      for (String repositoryNameDep : artifact.dependencies) {
+        Artifact targetNameDep = artifacts.get(repositoryNameDep);
         if (targetNameDep == null) {
           // our princess is in another castle!
-          bzlWriter.println(String.format("        # GMaven does not have %s" , repositoryNameDep));
+          bzlWriter.println(String.format("        # GMaven does not have %s", repositoryNameDep));
         } else {
-          bzlWriter.println(
-              String.format("        '%s',", repositoryNameToTargetName.get(repositoryNameDep)));
+          bzlWriter.println(String.format("        '%s',", targetNameDep.targetName()));
         }
       }
       bzlWriter.println("      ],");
@@ -120,7 +133,9 @@ public class GMavenToBazel {
   }
 
   private static Document parseUrl(URL url) throws Exception {
-    return documentBuilderFactory.newDocumentBuilder().parse(url.openStream());
+    try (InputStream in = new BufferedInputStream(url.openStream(), 8096)) {
+      return documentBuilderFactory.newDocumentBuilder().parse(in);
+    }
   }
 
   private static List<String> getGroupdIds(Document masterIndex) {
@@ -132,9 +147,13 @@ public class GMavenToBazel {
     return groupIds;
   }
 
-  private static Map<String, Set<String>> getArtifactsToVersions(String groupId)
-      throws Exception {
-    Document groupIndex = parseUrl(getGroupIndexUrl(groupId));
+  private static Map<String, Set<String>> getArtifactsToVersions(String groupId) {
+    Document groupIndex;
+    try {
+      groupIndex = parseUrl(getGroupIndexUrl(groupId));
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
     Map<String, Set<String>> result = new HashMap<>();
     NodeList artifactNodes = groupIndex.getDocumentElement().getChildNodes();
     for (int index = 1; index < artifactNodes.getLength(); index += 2) {
@@ -148,22 +167,23 @@ public class GMavenToBazel {
   }
 
   private static URL getGroupIndexUrl(String groupId) throws MalformedURLException {
-    return new URL(String.format(GROUP_INDEX_TEMPLATE_URL, groupId.replaceAll("\\.", "/")));
+    return new URL(String.format(GROUP_INDEX_TEMPLATE_URL, groupId.replace('.', '/')));
   }
 
-  private static URL getPomUrl(String groupId, String artifactId, String version) throws Exception {
+  private static URL getPomUrl(String groupId, String artifactId, String version) throws MalformedURLException {
     return new URL(
         String.format(
             POM_TEMPLATE_URL,
-            groupId.replaceAll("\\.", "/"),
+            groupId.replace('.', '/'),
             artifactId,
             version,
             artifactId,
             version));
   }
 
-  private static String getPackaging(Element element) throws Exception {
-    NodeList packaging = element.getElementsByTagName("packaging");
+  private static String getPackaging(Document pom) throws Exception {
+    Element project = (Element) pom.getElementsByTagName("project").item(0);
+    NodeList packaging = project.getElementsByTagName("packaging");
     return packaging.getLength() > 0 ? packaging.item(0).getTextContent() : "jar";
   }
 
@@ -190,21 +210,6 @@ public class GMavenToBazel {
     return escape(String.format("%s_%s_%s", groupId, artifactId, version));
   }
 
-  private static String getUrl(String groupId, String artifactId, String version, String packaging) {
-      return String.format(ARTIFACT_TEMPLATE_URL,
-                           groupId.replaceAll("\\.", "/"),
-                           artifactId,
-                           version,
-                           artifactId,
-                           version,
-                           packaging);
-  }
-
-  private static String getTargetName(
-      String groupId, String artifactId, String version, String packaging) {
-    return String.format("@%s//%s", getRepositoryName(groupId, artifactId, version), packaging);
-  }
-
   private static String escape(String string) {
     return string.replaceAll("[.-]", "_");
   }
@@ -214,5 +219,39 @@ public class GMavenToBazel {
     String raw = element.getElementsByTagName("version").item(0).getTextContent();
     raw = raw.replaceAll("\\[|]", "");
     return raw.split(",")[0];
+  }
+
+  static class Artifact {
+    String groupId;
+    String artifactId;
+    String version;
+    String packaging;
+    Set<String> dependencies;
+    String repositoryName;
+    Artifact(String groupId, String artifactId, String version, String packaging) {
+      this.groupId = groupId;
+      this.artifactId = artifactId;
+      this.version = version;
+      this.packaging = packaging;
+      this.repositoryName = getRepositoryName(groupId, artifactId, version);
+    }
+
+    String toArtifactUrl() {
+      return String.format(ARTIFACT_TEMPLATE_URL,
+                           groupId.replaceAll("\\.", "/"),
+                           artifactId,
+                           version,
+                           artifactId,
+                           version,
+                           packaging);
+    }
+
+    String targetName() {
+      return String.format("@%s//%s", repositoryName, packaging);
+    }
+
+    String rule() {
+      return PACKAGING_TO_RULE.get(packaging);
+    }
   }
 }